Refactor Reddit API Connection
