+++ backend/src/reddit/tests.rs
@@ -2,8 +2,8 @@
 mod tests {
     use lazy_static::lazy_static;
     use reqwest::{Client, ClientBuilder};
-    
-    use crate::reddit::{RedditConnection, RedditRequest};
+
+    use crate::reddit::{connection::RedditConnection, request::RedditRequest};
 
     lazy_static! {
         static ref HTTP: Client = ClientBuilder::new().user_agent("RMoods").build().unwrap();
